move core game + socketIO out of express server
